this is a function that compares genotype calls across two experiments

this is a function that compares genotype calls across two experiments

shared_genotypes(
  experiment_1_path,
  experiment_2_path,
  shared,
  experiment_1_name,
  experiment_2_name
)

Arguments

experiment_1_path
the file path to the first souporcell directory

experiment_2_path
the file path to the second souporcell directory

shared
numeric- the number of shared genotypes between the experiments

experiment_1_name
the name of experiment 1

experiment_2_name
the name of experiment 2

Value

a data frame of the shared genotype clusters between the two experiments

Examples

# \donttest{
#' shared_genotypes()
# }